backend/tests/test_main.py
"""Tests for Voice Dictation API."""
import pytest
from unittest.mock import patch, MagicMock, AsyncMock
from fastapi.testclient import TestClient
import sys
import os
sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))

from main import app

client = TestClient(app)


def test_health_check():
    response = client.get("/health")
    assert response.status_code == 200
    data = response.json()
    assert data["status"] == "ok"
    assert "whisper_loaded" in data


def test_transcribe_no_audio():
    # missing audio field
    response = client.post("/transcribe", json={})
    assert response.status_code in (400, 422)


def test_transcribe_invalid_language():
    response = client.post("/transcribe", json={
        "audio": "dGVzdA==",
        "language": "de"
    })
    assert response.status_code == 400


def test_transcribe_success():
    with patch("main.transcribe_audio", return_value="test transcribed text"), \
         patch("main.rewrite_with_ollama", new_callable=AsyncMock, return_value="Professional version."), \
         patch("main._whisper_model", MagicMock()):
        import base64
        audio_b64 = base64.b64encode(b"fake audio data").decode()
        response = client.post("/transcribe", json={
            "audio": audio_b64,
            "language": "pl",
            "use_local": True
        })
        # Should succeed or fail with whisper error (model not loaded in test env)
        assert response.status_code in (200, 400, 500)
        if response.status_code == 200:
            data = response.json()
            assert "original" in data
            assert "rewritten" in data
            assert data["success"] is True
